feat: use on_thread_create instead of on_thread_join
Why were we doing this in the first place? Also, shouldn't it only join every thread in r/igcse?

AFAIK, if you attempt to run a bot command in any thread it will automatically add the bot to the thread. Doing this makes sense for r/igcse so it can moderate etc.